Ensure correct timestamp handling on 32-bit systems

+++ src/V4l2Device.cpp
@@ -1389,7 +1389,7 @@ bool V4l2Device::get_frame ()
 
     // v4l2 timestamps contain seconds and microseconds
     // here they are converted to nanoseconds
-    statistics.capture_time_ns = (buf.timestamp.tv_sec * 1000 * 1000 * 1000) + (buf.timestamp.tv_usec * 1000);
+    statistics.capture_time_ns = ((long long)buf.timestamp.tv_sec * 1000 * 1000 * 1000) + (buf.timestamp.tv_usec * 1000);
     statistics.frame_count++;
     buffers.at(buf.index).buffer->set_statistics(statistics);
